+++ code/rtkConjugateGradientGetX_kPlusOneImageFilter.h
@@ -48,8 +48,8 @@ public:
   void SetXk(const TInputImage* Xk);
   void SetPk(const TInputImage* Pk);
 
-  itkGetMacro(Alphak, float)
-  itkSetMacro(Alphak, float)
+  itkGetMacro(Alphak, double)
+  itkSetMacro(Alphak, double)
 
   /** Typedefs for sub filters */
   typedef itk::AddImageFilter<TInputImage>                               AddFilterType;
@@ -70,7 +70,7 @@ protected:
 private:
   ConjugateGradientGetX_kPlusOneImageFilter(const Self &); //purposely not implemented
   void operator=(const Self &);  //purposely not implemented
-  float m_Alphak;
+  double m_Alphak;
 
   /** Pointers to sub filters */
   typename AddFilterType::Pointer       m_AddFilter;
